サーバー用ファイル名変更、データ応答確認用でもページ作成
