+++ app/templates/login.html
@@ -20,7 +20,7 @@
         <div class="login-tab" role="tabpanel">
             <a href="/top"><img class="wait-logo" src="static/img/wait-logo.png" alt="ワイトロゴ"></a>
             <div class="form-wrapper">
-                <form>
+                <form method="post" action="/mypage">
                     <div class="form-item">
                         <div class="form-item_login-id">
                             <label for="email">ログインID</label>
