+++ app/templates/user_regist.html
@@ -20,7 +20,7 @@
         <div class="user-regist-tab" role="tabpanel">
             <a href="/top"><img class="wait-logo" src="static/img/wait-logo.png" alt="ワイトロゴ"></a>
             <div class="form-wrapper">
-                <form id="myForm">
+                <form id="myForm" method="post" action="/user-info-regist">
                     <div class="form-item">
                         <!-- メールアドレス入力フォーム -->
                         <div class="form-item_user-regist">
@@ -39,7 +39,7 @@
                         <div class="form-item_repassword">
                             <label for="password">パスワード再入力</label>
                             <div class="form-item_repassword-input">
-                                <input id="re-password" type="password" name="password" required="required" placeholder="半角英数字8文字以上20文字以内"></input>
+                                <input id="re-password" type="password" name="repassword" required="required" placeholder="半角英数字8文字以上20文字以内"></input>
                                 <button id="re-view" onclick="onClickPassView('re-view', 're-password')" for="password" class="far fa-eye" type="button"></button>
                             </div>
                         </div>
